+++ frontend/.env.example
@@ -1,0 +1,1 @@
+BACKEND_API_URL = ""
