+++ frontend/src/utils/axios.utils.ts
@@ -1,0 +1,9 @@
+import axios from "axios";
+
+const axiosInstance = axios.create({
+  baseURL: import.meta.env.VITE_BACKEND_API_URL,
+  headers: {
+    "Content-type": "application/json; charset=UTF-8",
+  },
+});
+export default axiosInstance;
